+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Breaking
+
+- Routing types which have `PathParameters` use tuples rather than one of `NoPathParameters`, `OnePathParameter`, `ManyPathParameters`.
+
 ## [0.17.1] - 2025-11-07
 
 ### Fixed